conform default devise strategy to use params[:username]
Quite a few clients will only send :username and :password as opposed to :email and :password. So if a client is sending a :username, try to find a user based on the first  devise authentication symbol. A bit hacky but is a common enough practice we can deal with it in the default definition.