oPRO head now good for Rails 3 & 4
The biggest hurdle was concerning strong parameters which allows for mass assignment, i.e. using `update_attributes` or `create` and passing in a hash. Very useful for off the shelf CRUD apps, but not something oPRO was heavily using (or abusing). Instead of updating to the new Rails 4 way of doing things, we're taking out all mass assignment out of the app altogether. It does make code more verbose in a few places, but it also means that the code doesn't have to diverge between 3.x and 4.0.0 making maintaining and using easier.

While oPRO takes care of explicitly assigning attributes, users of oPRO in Rails4 will still need to appropriately use the appropriate sanitizing mechanism in user built controllers and users of oPRO in Rails3 need to sanitize their own models.