Fix Exception when Running in Rails 5.0
Rails 5.0 now raises an exception when `skip_before_action` (aka `skip_before_filter`) is called with a filter that does not exist (https://github.com/rails/rails/pull/19029). This causes opro to crash in some cases where before actions are skipped that may not exist. To resolve the issue, exceptions are explicitly silenced for skipped actions that may be defined elsewhere.